expense_apps/employee/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.10.1
Werkzeug==3.1.3

# Testing - Unit
pytest>=7.4.0
pytest-mock>=3.12.0
coverage>=7.3.0
allure-pytest>=2.13.0

# Testing - API
requests>=2.31.0

# Testing - E2E
behave>=1.2.6
selenium>=4.15.0
webdriver-manager>=4.0.0
allure-behave>=2.13.0
DateTime~=5.5
allure-python-commons~=2.15.3